requirements.txt
pandas==2.1.4
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
coverage==7.3.2
